requirements.txt
astroid==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pkg-resources==0.0.0
pycryptodome==3.8.2
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
